Use the `failure_derive` crate
Previously the code for `failure_derive` was vendored, with the
long-term goal of modifying it/enhancing it with some framework-specific
concerns.

However, `failure_derive` is a dependency of `failure` so we pull it in
anyway. Might as well use it, as opposed to cherry-picking a bunch of
upstream changes to proc macro code.