Remove vendored gumdrop; use upstream v0.5 crate
Originally abscissa vendored both `gumdrop` and `failure` to align
versions of syn and also prevent the possibility of hostile proc macros.

In hindsight, that was not a sustainable decision, and gumdrop in
particular has rather complex proc macros for which it would be nice to
receive upstream bugfixes.

This commit deletes all of the vendored gumdrop code, leaving
`abscissa_derive` an empty placeholder of a crate. But that's okay,
followup commits will replace some of the icky existing `macro_rules`
with fancy new proc macros.