abscissa_generator: Generate new apps with "abscissa new"
Adds support for generating new Abscissa apps from a built-in template.

Prior to this there was not a template for creating new Abscissa apps,
and wiring them up either involved looking through an existing app,
or intimate knowledge of how the framework works.

This adds a new `abscissa_generator` crate, and also adds a
`src/bin/abscissa` making this crate into a binary crate which contains
an `abscissa` command-line application (itself a proper Abscissa app!)
which will be available when `cargo install abscissa` is used.

The application template is implemented using Handlebars, and vendored
into the `abscissa_generator` crate (which is not included by
applications). The decision to use Handlebars was made after evaluating
a few different options for templating rust apps including
`cargo-generate` (which looks cool, but it seems like Abscissa could
benefit from something which both lives in the same git repo and is more
directly integrated with the framework, ala Rails)