Refactor Application and Config
This change makes Application-types the global state, as opposed to the
configuration. The configuration is now owned by the application.

This brings Abscissa a bit closer to an Entity-Component-System, where
the Application is the owner of all state, including the components, and
is thus able to obtain mutable references to its internal state.

Additionally this includes a fairly large refactor of application path
handing, adding a `Paths` associated type to `Application`, and
computing all paths in advance upon application startup.