thread: Wrappers for spawning threads
Adds an `abscissa::thread::Thread` type for threads managed by the
framework.

Each of these threads has a "kill switch" which can be used to request
that the thread exit. This is impl'd as an `AtomicBoolean` (and thus
does not have immediate effect, allowing threads to gracefully exit).

Each thread has a mandatory `abscissa::thread::Name` for identifying it.