testing: Add initial testing subsystem
Adds an `abscissa::testing` module with `CargoRunner` and `CmdRunner`
which is gated under the `testing` feature.

Enables the `testing` feature in the application boilerplate under
`dev-dependencies`.

Rewrites the existing app generator tests using this new functionality
(which is partially extracted from the original app generator tests),
exercising the test subsystem in an act of test inception.