generator: Add Default impl to generated config
This might be something the framework should require and leverage, so
apps always have a default configuration to start with which is replaced
with values from the config file once loaded.

For now though, the framework does not require the config types impl
`Default`, and it's merely a recommended best practice.