generator: Remove hashbrown dependency
handlebars 2.0.0 -> 2.0.1 upgraded `hashbrown` versions from 0.4 to 0.5
and in doing so broke the Abscissa v0.1 release, which now has an older
incompatible v0.4. I've opened an upstream issue about this here:

https://github.com/sunng87/handlebars-rust/issues/279

To avoid this in the future, this change removes the external
dependency. The implementation is a bit hacky, but gets the job done.