component: Use `generational-arena` for component storage
The `generational-arena` crate implements a generational index allocator
which provides a "weak ref" for addressing a particular component,
implemented in this PR in the form of `component::Handle`:

<https://github.com/fitzgen/generational-arena>

This also allows multiple simultaneous mutable references to components,
so this PR additionally introduces `Component::register_dependency`
which provides mutable refs to a component and its dependency for each
enumerated dependency. In principle this is equivalent to "dependency
injection".

This PR removes the previous downcasting functionality, which was never
tested and didn't seem to work out in practice. It should get added
back, but in a form that actually works, with tests.