component: Ensure registration will succeed before mutating
The previous implementation would encounter errors after adding
components to the `Arena`, and then try to clean up changes, which
accidentally deleted a key from the index in the process.

This error was captured in a test, then corrected by first ensuring that
registration would succeed before performing any state mutations.

Thanks to @dirvine for pointing out something was fishy.