Add `gimli-backtrace` Cargo feature
The `gimli` crate is an optional alternative backend for the `backtrace`
crate based on a pure-Rust DWARF parser as opposed to `libbacktrace`,
which is written in C. The long-term plan is to make `gimli` the default
backend for the `backtrace` crate (which will eventually be include in
`std` and is presently available on `nightly`).

For more information, see:

https://github.com/rust-lang/backtrace-rs/issues/189

This adds considerably to Cargo.toml as `gimli` has quite a few
dependencies, and because of that, it's left off-by-default, but with a
commented-out option to activate it in the template's Cargo.toml.