Remove `failure`
Since the time `failure` was written, `std::error::Error` has gained
many of the features it was originally useful for, most notably
`Error::source` for type erased error chains/cascades (with downcast
support to recover the concrete type, if desired), and on `nightly`,
backtrace support.

This commit attempts to remove `failure` in a way that largely preserves
the existing error handling, replacing `failure::Context` with an
internal `abscissa_core::error::Context` type which provides similar
functionality.

Additionally, it defines a `BoxError` alias for type-erased errors
based on ideas from this rust-internals thread:

https://internals.rust-lang.org/t/proposal-add-std-boxerror/10953/1

It's bounded as `Send + Sync + 'static` to help in multithreaded/async
contexts.

I'm not convinced this is good as a steady state, but rather could also
be a stepping stone towards moving to some more widely supported
library, as discussed in #144.