Completely replace `lazy_static` with `once_cell`
This eliminates all previous usages of `lazy_static`, including in the
new application template, with an `AppCell` type which is based on
`OnceCell`.

Unlike `lazy_static`, `OnceCell` does not require macros, and because of
that can be hidden within the framework as an implementation detail.

It's also on track for potential inclusion in the standard library,
which would allow us to eventually eliminate it as a dependency:

https://github.com/rust-lang/rfcs/pull/2788

Overall, the previous usage patterns for `lazy_static` were hacking
around the lack of a `OnceCell`-like API, so this is a net win.