README.md: update dependency list with changes in `tracing` dependencies
While I'm here, I took the liberty of updating the dependency list in
the README to reflect changes in transitive dependencies pulled in by
`tracing-subscriber` and `tracing`. Several dependencies were removed —
in particular, the dependency on `spin` — and a couple were added, but
I'm happy to note that the `tracing` update has a net-negative impact on
Abscissa's total number of dependencies.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>