core: remove signal handling (#419)
The current signal handling support is poorly thought out and not even
wired up by default.

Furthermore, it's getting in the way of other refactoring efforts, such
as moving to fine-grained locking, as the current implementation touches
many of the relevant data structures for that including the application
state and thread manager.

This commit removes it. We can revisit it after the other refactoring
efforts have been completed.