core: remove `generational-arena`
The original idea behind using `generational-arena` for component
storage was to support some kind of component reloading.

However, components aren't anywhere close to finished, and in the
meantime using something like `generational-arena` for this purpose is a
premature "optimization" which makes refactoring components more
difficult and introduces an unnecessary dependency.

This commit replaces `generational-arena` with a `Vec`, which is good
enough for now.

This is simple enough to swap out that the API is pretty much unchanged.