Switch CLI argument parsing to `clap` v3; MSRV 1.54
As discussed in #298, switches command-line argument processing from
`gumdrop` to `clap` v3.0.0-beta.4.

This release of `clap`, despite still being beta, has moved many of the
warnings around the stability of the API. They are also cutting scope of
an initial release upstream, which are good signs that they will get to
a stable release soon.