Update to `clap` v3.0 release
The final release of `clap` v3 is out.

This commit updates to it, and attempts to do things a bit more
idiomatically in regard to the upstream changes.

Namely, the re-exports have changed so only the `clap` crate itself is
re-exported, and the `clap::Parser` macro is used directly rather than
re-exported under the old `Clap` name.