derive: remove `darling` dependency
We're not really utilizing it much, and it adds considerably to compile
times.

The idea of declarative attribute parsers is a good one, however we're
not at the stage where it's worth the extra overhead.

The replacement code isn't great and misses a few key cases, but really
all of the component-related code is in need of an overhaul and can
perhaps be replaced with some existing well-maintained 3rd party library
for dependency injection.