Change `clap` bounds for `Command` from `Parser` to `FromArgMatches`
Clap has a separate `Subcommand` trait which is used instead of `Parser`
on enums which represent subcommands.

For that reason `Command` can't directly bound on `Parser` and still be
usable for subcommand enums.

Instead this commit changes `Command` to bound on the more general
`clap::FromArgMatches` trait and moves the `clap::Parser` bound to the
relevant methods for parsing toplevel arguments such as `parse_args` and
`parse_env_args`.